+++ modules/support/src/lib.rs
@@ -19,257 +19,29 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 #![allow(clippy::upper_case_acronyms)]
 
-use codec::{Decode, Encode, FullCodec};
+use codec::FullCodec;
 use frame_support::pallet_prelude::{DispatchClass, Pays, Weight};
-use primitives::{
-	evm::{CallInfo, EvmAddress},
-	task::TaskResult,
-	CurrencyId,
-};
-use scale_info::TypeInfo;
-use sp_core::H160;
+use primitives::{task::TaskResult, CurrencyId};
 use sp_runtime::{
-	traits::{AtLeast32BitUnsigned, CheckedDiv, MaybeSerializeDeserialize},
-	transaction_validity::TransactionValidityError,
-	DispatchError, DispatchResult, FixedU128, RuntimeDebug,
+	traits::CheckedDiv, transaction_validity::TransactionValidityError, DispatchError, DispatchResult, FixedU128,
 };
-use sp_std::{
-	cmp::{Eq, PartialEq},
-	fmt::Debug,
-	prelude::*,
-};
+use sp_std::prelude::*;
 
 use xcm::latest::prelude::*;
 
+pub mod dex;
+pub mod evm;
+pub mod honzon;
 pub mod mocks;
+
+pub use crate::dex::*;
+pub use crate::evm::*;
+pub use crate::honzon::*;
 
 pub type Price = FixedU128;
 pub type ExchangeRate = FixedU128;
 pub type Ratio = FixedU128;
 pub type Rate = FixedU128;
-
-pub trait RiskManager<AccountId, CurrencyId, Balance, DebitBalance> {
-	fn get_debit_value(currency_id: CurrencyId, debit_balance: DebitBalance) -> Balance;
-
-	fn check_position_valid(
-		currency_id: CurrencyId,
-		collateral_balance: Balance,
-		debit_balance: DebitBalance,
-		check_required_ratio: bool,
-	) -> DispatchResult;
-
-	fn check_debit_cap(currency_id: CurrencyId, total_debit_balance: DebitBalance) -> DispatchResult;
-}
-
-#[cfg(feature = "std")]
-impl<AccountId, CurrencyId, Balance: Default, DebitBalance> RiskManager<AccountId, CurrencyId, Balance, DebitBalance>
-	for ()
-{
-	fn get_debit_value(_currency_id: CurrencyId, _debit_balance: DebitBalance) -> Balance {
-		Default::default()
-	}
-
-	fn check_position_valid(
-		_currency_id: CurrencyId,
-		_collateral_balance: Balance,
-		_debit_balance: DebitBalance,
-		_check_required_ratio: bool,
-	) -> DispatchResult {
-		Ok(())
-	}
-
-	fn check_debit_cap(_currency_id: CurrencyId, _total_debit_balance: DebitBalance) -> DispatchResult {
-		Ok(())
-	}
-}
-
-pub trait AuctionManager<AccountId> {
-	type CurrencyId;
-	type Balance;
-	type AuctionId: FullCodec + Debug + Clone + Eq + PartialEq;
-
-	fn new_collateral_auction(
-		refund_recipient: &AccountId,
-		currency_id: Self::CurrencyId,
-		amount: Self::Balance,
-		target: Self::Balance,
-	) -> DispatchResult;
-	fn cancel_auction(id: Self::AuctionId) -> DispatchResult;
-	fn get_total_collateral_in_auction(id: Self::CurrencyId) -> Self::Balance;
-	fn get_total_target_in_auction() -> Self::Balance;
-}
-
-#[derive(RuntimeDebug, Encode, Decode, Clone, Copy, PartialEq, TypeInfo)]
-pub enum SwapLimit<Balance> {
-	/// use exact amount supply amount to swap. (exact_supply_amount, minimum_target_amount)
-	ExactSupply(Balance, Balance),
-	/// swap to get exact amount target. (maximum_supply_amount, exact_target_amount)
-	ExactTarget(Balance, Balance),
-}
-
-pub trait DEXManager<AccountId, CurrencyId, Balance> {
-	fn get_liquidity_pool(currency_id_a: CurrencyId, currency_id_b: CurrencyId) -> (Balance, Balance);
-
-	fn get_liquidity_token_address(currency_id_a: CurrencyId, currency_id_b: CurrencyId) -> Option<H160>;
-
-	fn get_swap_amount(path: &[CurrencyId], limit: SwapLimit<Balance>) -> Option<(Balance, Balance)>;
-
-	fn get_best_price_swap_path(
-		supply_currency_id: CurrencyId,
-		target_currency_id: CurrencyId,
-		limit: SwapLimit<Balance>,
-		alternative_path_joint_list: Vec<Vec<CurrencyId>>,
-	) -> Option<Vec<CurrencyId>>;
-
-	fn swap_with_specific_path(
-		who: &AccountId,
-		path: &[CurrencyId],
-		limit: SwapLimit<Balance>,
-	) -> sp_std::result::Result<(Balance, Balance), DispatchError>;
-
-	fn add_liquidity(
-		who: &AccountId,
-		currency_id_a: CurrencyId,
-		currency_id_b: CurrencyId,
-		max_amount_a: Balance,
-		max_amount_b: Balance,
-		min_share_increment: Balance,
-		stake_increment_share: bool,
-	) -> sp_std::result::Result<(Balance, Balance, Balance), DispatchError>;
-
-	fn remove_liquidity(
-		who: &AccountId,
-		currency_id_a: CurrencyId,
-		currency_id_b: CurrencyId,
-		remove_share: Balance,
-		min_withdrawn_a: Balance,
-		min_withdrawn_b: Balance,
-		by_unstake: bool,
-	) -> sp_std::result::Result<(Balance, Balance), DispatchError>;
-}
-
-#[cfg(feature = "std")]
-impl<AccountId, CurrencyId, Balance> DEXManager<AccountId, CurrencyId, Balance> for ()
-where
-	Balance: Default,
-{
-	fn get_liquidity_pool(_currency_id_a: CurrencyId, _currency_id_b: CurrencyId) -> (Balance, Balance) {
-		Default::default()
-	}
-
-	fn get_liquidity_token_address(_currency_id_a: CurrencyId, _currency_id_b: CurrencyId) -> Option<H160> {
-		Some(Default::default())
-	}
-
-	fn get_swap_amount(_path: &[CurrencyId], _limit: SwapLimit<Balance>) -> Option<(Balance, Balance)> {
-		Some(Default::default())
-	}
-
-	fn get_best_price_swap_path(
-		_supply_currency_id: CurrencyId,
-		_target_currency_id: CurrencyId,
-		_limit: SwapLimit<Balance>,
-		_alternative_path_joint_list: Vec<Vec<CurrencyId>>,
-	) -> Option<Vec<CurrencyId>> {
-		Some(Default::default())
-	}
-
-	fn swap_with_specific_path(
-		_who: &AccountId,
-		_path: &[CurrencyId],
-		_limit: SwapLimit<Balance>,
-	) -> sp_std::result::Result<(Balance, Balance), DispatchError> {
-		Ok(Default::default())
-	}
-
-	fn add_liquidity(
-		_who: &AccountId,
-		_currency_id_a: CurrencyId,
-		_currency_id_b: CurrencyId,
-		_max_amount_a: Balance,
-		_max_amount_b: Balance,
-		_min_share_increment: Balance,
-		_stake_increment_share: bool,
-	) -> sp_std::result::Result<(Balance, Balance, Balance), DispatchError> {
-		Ok(Default::default())
-	}
-
-	fn remove_liquidity(
-		_who: &AccountId,
-		_currency_id_a: CurrencyId,
-		_currency_id_b: CurrencyId,
-		_remove_share: Balance,
-		_min_withdrawn_a: Balance,
-		_min_withdrawn_b: Balance,
-		_by_unstake: bool,
-	) -> sp_std::result::Result<(Balance, Balance), DispatchError> {
-		Ok(Default::default())
-	}
-}
-
-/// An abstraction of cdp treasury for Honzon Protocol.
-pub trait CDPTreasury<AccountId> {
-	type Balance;
-	type CurrencyId;
-
-	/// get surplus amount of cdp treasury
-	fn get_surplus_pool() -> Self::Balance;
-
-	/// get debit amount of cdp treasury
-	fn get_debit_pool() -> Self::Balance;
-
-	/// get collateral assets amount of cdp treasury
-	fn get_total_collaterals(id: Self::CurrencyId) -> Self::Balance;
-
-	/// calculate the proportion of specific debit amount for the whole system
-	fn get_debit_proportion(amount: Self::Balance) -> Ratio;
-
-	/// issue debit for cdp treasury
-	fn on_system_debit(amount: Self::Balance) -> DispatchResult;
-
-	/// issue surplus(stable currency) for cdp treasury
-	fn on_system_surplus(amount: Self::Balance) -> DispatchResult;
-
-	/// issue debit to `who`
-	/// if backed flag is true, means the debit to issue is backed on some
-	/// assets, otherwise will increase same amount of debit to system debit.
-	fn issue_debit(who: &AccountId, debit: Self::Balance, backed: bool) -> DispatchResult;
-
-	/// burn debit(stable currency) of `who`
-	fn burn_debit(who: &AccountId, debit: Self::Balance) -> DispatchResult;
-
-	/// deposit surplus(stable currency) to cdp treasury by `from`
-	fn deposit_surplus(from: &AccountId, surplus: Self::Balance) -> DispatchResult;
-
-	/// deposit collateral assets to cdp treasury by `who`
-	fn deposit_collateral(from: &AccountId, currency_id: Self::CurrencyId, amount: Self::Balance) -> DispatchResult;
-
-	/// withdraw collateral assets of cdp treasury to `who`
-	fn withdraw_collateral(to: &AccountId, currency_id: Self::CurrencyId, amount: Self::Balance) -> DispatchResult;
-}
-
-pub trait CDPTreasuryExtended<AccountId>: CDPTreasury<AccountId> {
-	fn swap_collateral_to_stable(
-		currency_id: Self::CurrencyId,
-		limit: SwapLimit<Self::Balance>,
-		collateral_in_auction: bool,
-	) -> sp_std::result::Result<(Self::Balance, Self::Balance), DispatchError>;
-
-	fn create_collateral_auctions(
-		currency_id: Self::CurrencyId,
-		amount: Self::Balance,
-		target: Self::Balance,
-		refund_receiver: AccountId,
-		splited: bool,
-	) -> sp_std::result::Result<u32, DispatchError>;
-
-	fn remove_liquidity_for_lp_collateral(
-		currency_id: Self::CurrencyId,
-		amount: Self::Balance,
-	) -> sp_std::result::Result<(Self::Balance, Self::Balance), DispatchError>;
-
-	fn max_auction() -> u32;
-}
 
 pub trait PriceProvider<CurrencyId> {
 	fn get_price(currency_id: CurrencyId) -> Option<Price>;
@@ -295,10 +67,6 @@ pub trait ExchangeRateProvider {
 	fn get_exchange_rate() -> ExchangeRate;
 }
 
-pub trait EmergencyShutdown {
-	fn is_shutdown() -> bool;
-}
-
 pub trait DEXIncentives<AccountId, CurrencyId, Balance> {
 	fn do_deposit_dex_share(who: &AccountId, lp_currency_id: CurrencyId, amount: Balance) -> DispatchResult;
 	fn do_withdraw_dex_share(who: &AccountId, lp_currency_id: CurrencyId, amount: Balance) -> DispatchResult;
@@ -313,119 +81,6 @@ impl<AccountId, CurrencyId, Balance> DEXIncentives<AccountId, CurrencyId, Balanc
 	fn do_withdraw_dex_share(_: &AccountId, _: CurrencyId, _: Balance) -> DispatchResult {
 		Ok(())
 	}
-}
-
-/// Return true if the call of EVM precompile contract is allowed.
-pub trait PrecompileCallerFilter {
-	fn is_allowed(caller: H160) -> bool;
-}
-
-/// An abstraction of EVM for EVMBridge
-pub trait EVM<AccountId> {
-	type Balance: AtLeast32BitUnsigned + Copy + MaybeSerializeDeserialize + Default;
-
-	fn execute(
-		context: InvokeContext,
-		input: Vec<u8>,
-		value: Self::Balance,
-		gas_limit: u64,
-		storage_limit: u32,
-		mode: ExecutionMode,
-	) -> Result<CallInfo, sp_runtime::DispatchError>;
-
-	/// Get the real origin account and charge storage rent from the origin.
-	fn get_origin() -> Option<AccountId>;
-	/// Provide a method to set origin for `on_initialize`
-	fn set_origin(origin: AccountId);
-}
-
-#[derive(Encode, Decode, Eq, PartialEq, Copy, Clone, RuntimeDebug)]
-pub enum ExecutionMode {
-	Execute,
-	/// Discard any state changes
-	View,
-	/// Also discard any state changes and use estimate gas mode for evm config
-	EstimateGas,
-}
-
-#[derive(Encode, Decode, Eq, PartialEq, Copy, Clone, RuntimeDebug)]
-pub struct InvokeContext {
-	pub contract: EvmAddress,
-	/// similar to msg.sender
-	pub sender: EvmAddress,
-	/// similar to tx.origin
-	pub origin: EvmAddress,
-}
-
-/// An abstraction of EVMBridge
-pub trait EVMBridge<AccountId, Balance> {
-	/// Execute ERC20.name() to read token name from ERC20 contract
-	fn name(context: InvokeContext) -> Result<Vec<u8>, DispatchError>;
-	/// Execute ERC20.symbol() to read token symbol from ERC20 contract
-	fn symbol(context: InvokeContext) -> Result<Vec<u8>, DispatchError>;
-	/// Execute ERC20.decimals() to read token decimals from ERC20 contract
-	fn decimals(context: InvokeContext) -> Result<u8, DispatchError>;
-	/// Execute ERC20.totalSupply() to read total supply from ERC20 contract
-	fn total_supply(context: InvokeContext) -> Result<Balance, DispatchError>;
-	/// Execute ERC20.balanceOf(address) to read balance of address from ERC20
-	/// contract
-	fn balance_of(context: InvokeContext, address: EvmAddress) -> Result<Balance, DispatchError>;
-	/// Execute ERC20.transfer(address, uint256) to transfer value to `to`
-	fn transfer(context: InvokeContext, to: EvmAddress, value: Balance) -> DispatchResult;
-	/// Get the real origin account and charge storage rent from the origin.
-	fn get_origin() -> Option<AccountId>;
-	/// Provide a method to set origin for `on_initialize`
-	fn set_origin(origin: AccountId);
-}
-
-#[cfg(feature = "std")]
-impl<AccountId, Balance: Default> EVMBridge<AccountId, Balance> for () {
-	fn name(_context: InvokeContext) -> Result<Vec<u8>, DispatchError> {
-		Err(DispatchError::Other("unimplemented evm bridge"))
-	}
-	fn symbol(_context: InvokeContext) -> Result<Vec<u8>, DispatchError> {
-		Err(DispatchError::Other("unimplemented evm bridge"))
-	}
-	fn decimals(_context: InvokeContext) -> Result<u8, DispatchError> {
-		Err(DispatchError::Other("unimplemented evm bridge"))
-	}
-	fn total_supply(_context: InvokeContext) -> Result<Balance, DispatchError> {
-		Err(DispatchError::Other("unimplemented evm bridge"))
-	}
-	fn balance_of(_context: InvokeContext, _address: EvmAddress) -> Result<Balance, DispatchError> {
-		Err(DispatchError::Other("unimplemented evm bridge"))
-	}
-	fn transfer(_context: InvokeContext, _to: EvmAddress, _value: Balance) -> DispatchResult {
-		Err(DispatchError::Other("unimplemented evm bridge"))
-	}
-	fn get_origin() -> Option<AccountId> {
-		None
-	}
-	fn set_origin(_origin: AccountId) {}
-}
-
-/// An abstraction of EVMManager
-pub trait EVMManager<AccountId, Balance> {
-	/// Query the constants `NewContractExtraBytes` value from evm module.
-	fn query_new_contract_extra_bytes() -> u32;
-	/// Query the constants `StorageDepositPerByte` value from evm module.
-	fn query_storage_deposit_per_byte() -> Balance;
-	/// Query the maintainer address from the ERC20 contract.
-	fn query_maintainer(contract: H160) -> Result<H160, DispatchError>;
-	/// Query the constants `DeveloperDeposit` value from evm module.
-	fn query_developer_deposit() -> Balance;
-	/// Query the constants `PublicationFee` value from evm module.
-	fn query_publication_fee() -> Balance;
-	/// Transfer the maintainer of the contract address.
-	fn transfer_maintainer(from: AccountId, contract: H160, new_maintainer: H160) -> DispatchResult;
-	/// Publish contract
-	fn publish_contract_precompile(who: AccountId, contract: H160) -> DispatchResult;
-	/// Query the developer status of an account
-	fn query_developer_status(who: AccountId) -> bool;
-	/// Enable developer mode
-	fn enable_account_contract_development(who: AccountId) -> DispatchResult;
-	/// Disable developer mode
-	fn disable_account_contract_development(who: AccountId) -> DispatchResult;
 }
 
 pub trait TransactionPayment<AccountId, Balance, NegativeImbalance> {
@@ -482,91 +137,6 @@ impl<AccountId, Balance: Default + Copy, NegativeImbalance: Imbalance<Balance>>
 		_class: DispatchClass,
 	) -> Result<(), TransactionValidityError> {
 		Ok(())
-	}
-}
-
-pub trait Contains<T> {
-	fn contains(t: &T) -> bool;
-}
-
-/// A mapping between `AccountId` and `EvmAddress`.
-pub trait AddressMapping<AccountId> {
-	/// Returns the AccountId used go generate the given EvmAddress.
-	fn get_account_id(evm: &EvmAddress) -> AccountId;
-	/// Returns the EvmAddress associated with a given AccountId or the
-	/// underlying EvmAddress of the AccountId.
-	/// Returns None if there is no EvmAddress associated with the AccountId
-	/// and there is no underlying EvmAddress in the AccountId.
-	fn get_evm_address(account_id: &AccountId) -> Option<EvmAddress>;
-	/// Returns the EVM address associated with an account ID and generates an
-	/// account mapping if no association exists.
-	fn get_or_create_evm_address(account_id: &AccountId) -> EvmAddress;
-	/// Returns the default EVM address associated with an account ID.
-	fn get_default_evm_address(account_id: &AccountId) -> EvmAddress;
-	/// Returns true if a given AccountId is associated with a given EvmAddress
-	/// and false if is not.
-	fn is_linked(account_id: &AccountId, evm: &EvmAddress) -> bool;
-}
-
-/// A mapping between AssetId and AssetMetadata.
-pub trait AssetIdMapping<StableAssetPoolId, ForeignAssetId, MultiLocation, AssetMetadata> {
-	/// Returns the AssetMetadata associated with a given contract address.
-	fn get_erc20_asset_metadata(contract: EvmAddress) -> Option<AssetMetadata>;
-	/// Returns the AssetMetadata associated with a given StableAssetPoolId.
-	fn get_stable_asset_metadata(stable_asset_id: StableAssetPoolId) -> Option<AssetMetadata>;
-	/// Returns the AssetMetadata associated with a given ForeignAssetId.
-	fn get_foreign_asset_metadata(foreign_asset_id: ForeignAssetId) -> Option<AssetMetadata>;
-	/// Returns the MultiLocation associated with a given ForeignAssetId.
-	fn get_multi_location(foreign_asset_id: ForeignAssetId) -> Option<MultiLocation>;
-	/// Returns the CurrencyId associated with a given MultiLocation.
-	fn get_currency_id(multi_location: MultiLocation) -> Option<CurrencyId>;
-}
-
-/// A mapping between u32 and Erc20 address.
-/// provide a way to encode/decode for CurrencyId;
-pub trait Erc20InfoMapping {
-	/// Returns the name associated with a given CurrencyId.
-	/// If CurrencyId is CurrencyId::DexShare and contain DexShare::Erc20,
-	/// the EvmAddress must have been mapped.
-	fn name(currency_id: CurrencyId) -> Option<Vec<u8>>;
-	/// Returns the symbol associated with a given CurrencyId.
-	/// If CurrencyId is CurrencyId::DexShare and contain DexShare::Erc20,
-	/// the EvmAddress must have been mapped.
-	fn symbol(currency_id: CurrencyId) -> Option<Vec<u8>>;
-	/// Returns the decimals associated with a given CurrencyId.
-	/// If CurrencyId is CurrencyId::DexShare and contain DexShare::Erc20,
-	/// the EvmAddress must have been mapped.
-	fn decimals(currency_id: CurrencyId) -> Option<u8>;
-	/// Encode the CurrencyId to EvmAddress.
-	/// If is CurrencyId::DexShare and contain DexShare::Erc20,
-	/// will use the u32 to get the DexShare::Erc20 from the mapping.
-	fn encode_evm_address(v: CurrencyId) -> Option<EvmAddress>;
-	/// Decode the CurrencyId from EvmAddress.
-	/// If is CurrencyId::DexShare and contain DexShare::Erc20,
-	/// will use the u32 to get the DexShare::Erc20 from the mapping.
-	fn decode_evm_address(v: EvmAddress) -> Option<CurrencyId>;
-}
-
-#[cfg(feature = "std")]
-impl Erc20InfoMapping for () {
-	fn name(_currency_id: CurrencyId) -> Option<Vec<u8>> {
-		None
-	}
-
-	fn symbol(_currency_id: CurrencyId) -> Option<Vec<u8>> {
-		None
-	}
-
-	fn decimals(_currency_id: CurrencyId) -> Option<u8> {
-		None
-	}
-
-	fn encode_evm_address(_v: CurrencyId) -> Option<EvmAddress> {
-		None
-	}
-
-	fn decode_evm_address(_v: EvmAddress) -> Option<CurrencyId> {
-		None
 	}
 }
